Pattern info using lookup table
To correctly display mods that jump back and forth in a pattern the
entire pattern info view is now updated in each frame. Using JQuery to
lookup the table entries to change is too slow when fast songs are
played so a lookup table is used that references the various pattern
data cells.

Also PositionJump and PatternBreak effects now function correctly.